adição de add OrgUnit

+++ MVC-2020-Database/DataModels/OrgUnit_Classification.cs
@@ -19,6 +19,6 @@ namespace MVC_2020_Database.DataModels
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
 
-        public int Fraction { get; set; }
+        public float Fraction { get; set; }
     }
 }
